add rename and stats components
