+++ dammit/cli.py
@@ -1,0 +1,13 @@
+# Copyright (C) 2015-2019 Camille Scott
+# All rights reserved.
+#
+# This software may be modified and distributed under the terms
+# of the BSD license.  See the LICENSE file for details.
+
+import click
+
+@click.group()
+def component():
+    pass
+
+from .components import *
